feat: v1.3.4 — customer messaging, OCR overhaul, delivery team reviews, signed release
- Enhanced OCR extraction: brand detection (30+ brands), serial/model numbers,
  per-item service flags, alt phone numbers, store phone suppression
- Customer messaging system: follow-up SMS on delivery completion with
  3 editable message templates and smart placeholders
- Delivery team branding: team name tag in all messages and suggested reviews
- Pre-written 5-star Google review template customers can copy/paste
- Auto URL shortening via TinyURL for cleaner SMS messages
- Visual delivery completion: greyed-out cards at 45% opacity
- Service type badges on invoice cards
- Manual extraction crash fixes and OCR region fallback
- Room DB migration 9→10 (altPhone column)
- Signed release build configuration
- Version bump to 1.3.4 (versionCode 134)

+++ app/src/main/java/com/mobileinvoice/ocr/InvoiceTemplate.java
@@ -29,10 +29,10 @@ public class InvoiceTemplate {
         public String type;
 
         public Rect toPixelRect(int imageWidth, int imageHeight, Grid grid) {
-            int left = (int) ((this.coords.x1 / grid.cols) * imageWidth);
-            int top = (int) ((this.coords.y1 / grid.rows) * imageHeight);
-            int right = (int) ((this.coords.x2 / grid.cols) * imageWidth);
-            int bottom = (int) ((this.coords.y2 / grid.rows) * imageHeight);
+            int left = (this.coords.x1 / grid.cols) * imageWidth;
+            int top = (this.coords.y1 / grid.rows) * imageHeight;
+            int right = (this.coords.x2 / grid.cols) * imageWidth;
+            int bottom = (this.coords.y2 / grid.rows) * imageHeight;
             return new Rect(left, top, right, bottom);
         }
     }
